+++ llvm/docs/ReleaseNotes.rst
@@ -190,6 +190,9 @@ Changes to the LLVM tools
 Changes to LLDB
 ---------------------------------
 
+* Register field information is now provided on AArch64 FreeBSD for live
+  processes and core files (previously only provided on AArch64 Linux).
+
 Changes to Sanitizers
 ---------------------